Restructured bulk prenotification request handler tests.
