+++ src/EA.Iws.RequestHandlers.Tests.Unit/EA.Iws.RequestHandlers.Tests.Unit.csproj
@@ -167,23 +167,23 @@
     <Compile Include="NotificationMovements\BulkReceiptRecovery\ReceiptRecoveryReceiptDateFormatRuleTests.cs" />
     <Compile Include="NotificationMovements\BulkReceiptRecovery\ReceiptRecoveryReceiptDateValidationRuleTests.cs" />
     <Compile Include="NotificationMovements\BulkUpload\GetBulkUploadTemplateHandlerTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PerformBulkUploadContentValidationHandlerTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationContentDateHistoricRuleTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationContentDuplicateShipmentNumberRuleTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationDateFormatRuleTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationDateFutureRuleTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationInvalidPackagingTypeRuleTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationInvalidShipmentNumbersRuleTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationNotificationNumberRuleTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationOnlyNewShipmentsRuleTests.cs.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationQuantityExceededRuleTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationQuantityNumericTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationQuantityPrecisionForCubicMetresAndTonnesRuleTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationQuantityPrecisionForLitresAndKilogramsRuleTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationContentConsentValidityRuleTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationContentQuantityUnitRuleTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationContentThreeWorkingDaysConsentDateRuleTests.cs" />
-    <Compile Include="NotificationMovements\BulkUpload\PrenotificationContentThreeWorkingDaysRuleTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PerformBulkUploadContentValidationHandlerTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationContentDateHistoricRuleTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationContentDuplicateShipmentNumberRuleTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationDateFormatRuleTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationDateFutureRuleTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationInvalidPackagingTypeRuleTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationInvalidShipmentNumbersRuleTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationNotificationNumberRuleTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationOnlyNewShipmentsRuleTests.cs.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationQuantityExceededRuleTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationQuantityNumericTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationQuantityPrecisionForCubicMetresAndTonnesRuleTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationQuantityPrecisionForLitresAndKilogramsRuleTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationContentConsentValidityRuleTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationContentQuantityUnitRuleTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationContentThreeWorkingDaysConsentDateRuleTests.cs" />
+    <Compile Include="NotificationMovements\BulkPrenotification\PrenotificationContentThreeWorkingDaysRuleTests.cs" />
     <Compile Include="NotificationMovements\Capture\EnsureMovementNumberAvailableHandlerTests.cs" />
     <Compile Include="NotificationMovements\Create\CreateMovementCarriersHandlerTests.cs" />
     <Compile Include="NotificationMovements\BulkReceiptRecovery\PerformReceiptRecoveryContentValidationHandlerTests.cs" />
