+++ src/EA.Iws.RequestHandlers.Tests.Unit/NotificationMovements/BulkPrenotification/PrenotificationQuantityNumericTests.cs
@@ -1,57 +1,56 @@
-﻿namespace EA.Iws.RequestHandlers.Tests.Unit.NotificationMovements.BulkUpload
+﻿namespace EA.Iws.RequestHandlers.Tests.Unit.NotificationMovements.BulkPrenotification
 {
     using System;
     using System.Collections.Generic;
     using System.Threading.Tasks;
     using Core.Movement.BulkPrenotification;
     using Core.Rules;
-    using Prsd.Core;
     using RequestHandlers.NotificationMovements.BulkPrenotification;
     using Xunit;
 
-    public class PrenotificationDateFormatRuleTests
+    public class PrenotificationQuantityNumericTests
     {
-        private readonly PrenotificationDateFormatRule rule;
+        private readonly PrenotificationQuantityNumericRule rule;
 
-        public PrenotificationDateFormatRuleTests()
+        public PrenotificationQuantityNumericTests()
         {
-            rule = new PrenotificationDateFormatRule();
+            rule = new PrenotificationQuantityNumericRule();
         }
 
         [Fact]
-        public async Task DateRule_Success()
+        public async Task PrenotificationQuantityNumericRule_ValidData_Error()
         {
             var movements = new List<PrenotificationMovement>()
             {
                 new PrenotificationMovement()
                 {
                     ShipmentNumber = 1,
-                    ActualDateOfShipment = SystemTime.UtcNow
+                    Quantity = null
                 }
             };
 
             var result = await rule.GetResult(movements, Guid.NewGuid());
 
-            Assert.Equal(PrenotificationContentRules.InvalidDateFormat, result.Rule);
-            Assert.Equal(MessageLevel.Success, result.MessageLevel);
-        }
-
-        [Fact]
-        public async Task DateRule_Error()
-        {
-            var movements = new List<PrenotificationMovement>()
-            {
-                new PrenotificationMovement()
-                {
-                    ShipmentNumber = 1,
-                    ActualDateOfShipment = null
-                }
-            };
-
-            var result = await rule.GetResult(movements, Guid.NewGuid());
-
-            Assert.Equal(PrenotificationContentRules.InvalidDateFormat, result.Rule);
+            Assert.Equal(PrenotificationContentRules.QuantityNumeric, result.Rule);
             Assert.Equal(MessageLevel.Error, result.MessageLevel);
+        }
+
+        [Fact]
+        public async Task PrenotificationQuantityNumericRule_InvalidData_Error()
+        {
+            var movements = new List<PrenotificationMovement>()
+            {
+                new PrenotificationMovement()
+                {
+                    ShipmentNumber = 1,
+                    Quantity = 10m
+                }
+            };
+
+            var result = await rule.GetResult(movements, Guid.NewGuid());
+
+            Assert.Equal(PrenotificationContentRules.QuantityNumeric, result.Rule);
+            Assert.Equal(MessageLevel.Success, result.MessageLevel);
         }
     }
 }
